Module Irmin_unix.FS

File system backends, using bin_prot.

File-system Store

module Append_only : Irmin.APPEND_ONLY_STORE_MAKER

Append-only store maker.

module Atomic_write : Irmin.ATOMIC_WRITE_STORE_MAKER

Atomic-write store maker.

module Make : Irmin.S_MAKER

Irmin store maker.

module KV : Irmin.KV_MAKER

Irmin store make, where only the Contents have to be specified: branches are strings and paths are string lists.

module Append_only_ext (C : Irmin_fs.Config) : Irmin.APPEND_ONLY_STORE_MAKER

Append-only store maker, with control over the filenames shapes.

module Atomic_write_ext (C : Irmin_fs.Config) : Irmin.ATOMIC_WRITE_STORE_MAKER

Read-write store maker, with control over the filename shapes.

module Make_ext (Obj : Irmin_fs.Config) (Ref : Irmin_fs.Config) : Irmin.S_MAKER

Irmin store maker, with control over the filename shapes.