Module 1-S.Info

type author = string

val author_t : author Irmin.Type.t

type message = string

val message_t : message Irmin.Type.t

type t = info

The type for commit info.

val t : t Irmin.Type.t

val v : ?author:author -> ?message:message -> int64 -> t

Create a new commit info.

val date : t -> int64

date t is t's commit date.

The date provided by the user when calling the v function. Rounding Unix.gettimeofday () (when available) is a good value for such date. On more esoteric platforms, any monotonic counter is a fine value as well. On the Git backend, the date is translated into the commit Date field and is expected to be the number of POSIX seconds (thus not counting leap seconds) since the Epoch.

val author : t -> author

author t is t's commit author.

The author identifies the entity (human, unikernel, process, thread, etc) performing an operation. For the Git backend, this will be directly translated into the Author field.

val message : t -> message

message t is t's commit message.

val empty : t

The empty commit info.

Info Functions

type f = unit -> t

Alias for functions which can build commit info.

val none : f

The empty info function. none () is empty