Module Irmin_unix.Resolver

Irmin store resolver.

val global_option_section : string

val branch : string option Cmdliner.Term.t

module Hash : sig ... end

type hash = Hash.t

module Contents : sig ... end

type contents = Contents.t

Global Configuration

module Store : sig ... end

Stores

val load_config : ?root:string -> ?config_path:string -> ?store:string -> ?hash:hash -> ?contents:string -> unit -> Store.t * Irmin.config

Load config file from disk

config_path can be used to specify the location of a configuration file.

root is used to specify the path of the store.

The values provided for store, hash and contents will be used by default if no other value is found in the config file

type store =

| S : 'a Store.Impl.t * 'a Lwt.t * Store.remote_fn option -> store

val store : unit -> store Cmdliner.Term.t

Parse the command-line arguments and then the config file.

type Irmin.remote +=

| R of Cohttp.Header.t option * string

val remote : unit -> (store * Irmin.remote Lwt.t) Cmdliner.Term.t

Parse a remote store location.