Module Private.Repo

Private repositories.

type t = repo

type 'a t

val v : Irmin__.S.config -> Irmin__.Import.read t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

val contents_t : t -> Irmin__.Import.read Contents.t

val node_t : t -> Irmin__.Import.read Node.t

val commit_t : t -> Irmin__.Import.read Commit.t

val branch_t : t -> Branch.t

val batch : t -> (Irmin__.Import.read_write Contents.t -> Irmin__.Import.read_write Node.t -> Irmin__.Import.read_write Commit.t -> 'a Lwt.t) -> 'a Lwt.t