Module Private.Contents

Private content store.

include Irmin__.S.CONTENT_ADDRESSABLE_STORE

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin__.Import.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin__.Import.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

include Irmin__.S.Store_properties.CLEARABLE with type 'a t := 'a t

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

val merge : [> Irmin__.Import.read_write ] t -> key option Irmin.Merge.t

merge t lifts the merge functions defined on contents values to contents key. The merge function will: (i) read the values associated with the given keys, (ii) use the merge function defined on values and (iii) write the resulting values into the store to get the resulting key. See Irmin__.Contents_intf.Contents.S.merge.

If any of these operations fail, return `Conflict.

module Key : Irmin.Hash.TYPED with type t = key and type value = value

Key provides base functions for user-defined contents keys.

module Val : Irmin__.Contents_intf.S with type t = value

Val provides base functions for user-defined contents values.