Module Http.Server

Parameters

S : Irmin.S

Signature

type repo = S.Repo.t

The type for Irmin repository.

type t = Cohttp_lwt_unix.Server.t

The type for HTTP configuration.

val v : ?⁠strict:bool -> repo -> t

v repo returns the configuration for a server serving the contents of repo. If strict is set, incoming connections will fail if they do not have the right X-IrminVersion headers.