Module Irmin_unix

Unix backends

This module provides Irmin backends for Unix applications. The currently supported backends are:

An in-memory store, internally using hash tables.

A file-system backend, using bin_prot to serialize internal structures.

A fully compatible, bidirectional encoding of Irmin into Git. You can view and edit your store using both the library and your usual Git tools.

The HTTP clients and servers provides a high-level REST API, with 1 RTT for the private and public functions.

module Info = Info.Make

val info : ?author:string -> ( 'a, Stdlib.Format.formatter, unit, unit -> Irmin.Info.default ) Stdlib.format4 -> 'a

info fmt () creates a fresh commit info, with the date set to Unix.gettimeoday () and the author built using Unix.gethostname() and Unix.getpid() if author is not provided.

module FS : sig ... end

File system backends, using bin_prot.

module Git : sig ... end

Bidirectional Git backends.

module Http : sig ... end

REST (over HTTP) backend..

val set_listen_dir_hook : unit -> unit

Install Irmin_watcher.hook as the listen hook for watching changes in directories.

module Graphql : sig ... end

module Cli : sig ... end

CLI commands.

module Resolver : sig ... end

Irmin store resolver.