Module Irmin_unix.FS

File system backends, using bin_prot.

File-system Store

module Append_only : Irmin.Append_only.Maker

Append-only store maker.

module Atomic_write : Irmin.Atomic_write.Maker

Atomic-write store maker.

Irmin store maker.

include Irmin.Maker

type endpoint

include Irmin.Key.Store_spec.S

type ('h, 'v) contents_key

type 'h node_key

type 'h commit_key

module Make : functor (Schema : Irmin.Schema.S) -> Irmin__.Store_intf.S_generic_key with module Schema = Schema and type Backend.Remote.endpoint = endpoint and type contents_key = (Schema.Hash.t, Schema.Contents.t) contents_key and type node_key = Schema.Hash.t node_key and type commit_key = Schema.Hash.t commit_key

module KV : Irmin.KV_maker

Irmin store make, where only the Contents have to be specified: branches are strings and paths are string lists.

module Append_only_ext : functor (C : Irmin_fs.Config) -> Irmin.Append_only.Maker

Append-only store maker, with control over the filenames shapes.

module Atomic_write_ext : functor (C : Irmin_fs.Config) -> Irmin.Atomic_write.Maker

Read-write store maker, with control over the filename shapes.

module Maker_ext : functor (Obj : Irmin_fs.Config) -> functor (Ref : Irmin_fs.Config) -> Irmin.Maker

Irmin store maker, with control over the filename shapes.