Module Irmin_unix.FS

File system backends, using bin_prot.

File-system Store

module Append_only : Irmin.Append_only.Maker

Append-only store maker.

module Atomic_write : Irmin.Atomic_write.Maker

Atomic-write store maker.

Irmin store maker.

include Irmin.Maker

type endpoint

include Irmin.Key.Store_spec.S with type ('h, _) contents_key = 'h with type 'h node_key = 'h with type 'h commit_key = 'h

type ('h, _) contents_key = 'h

type 'h node_key = 'h

type 'h commit_key = 'h

module Make (Schema : Irmin.Schema.S) : sig ... end

module KV : Irmin.KV_maker

Irmin store make, where only the Contents have to be specified: branches are strings and paths are string lists.

module Append_only_ext (C : Irmin_fs.Config) : Irmin.Append_only.Maker

Append-only store maker, with control over the filenames shapes.

module Atomic_write_ext (C : Irmin_fs.Config) : Irmin.Atomic_write.Maker

Read-write store maker, with control over the filename shapes.

module Maker_ext (Obj : Irmin_fs.Config) (Ref : Irmin_fs.Config) : Irmin.Maker

Irmin store maker, with control over the filename shapes.